requirements.txt
requests==2.32.3
dask==2023.6.0
easyocr==1.7.1
mercantile==1.2.1
numpy==1.24.0
pandas==2.2.2
Pillow==11.0.0
rasterio==1.3.9
Requests==2.32.3
selenium==4.23.1
undetected_chromedriver==3.5.5
Shapely==2.0.6
xarray==2024.2.0
hvplot==0.9.1
cryptography==41.0.3
odc-stac==0.3.9
pystac_client==0.7.2
planetary_computer==1.0.0
paddlepaddle==2.6.1
paddleocr==2.8.1
rioxarray==0.15.3
xee==0.0.9
ipython==8.12.3
geopandas==0.14.4
